fix | refacto: reduce code duplication
